pyproject.toml
[project]
name = "openhands-eval"
version = "0.1.0"
description = "OpenHands agent evaluation with Qwen3 Coder via OpenRouter"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "litellm>=1.0.0",
    "openhands-sdk",
    "openhands-tools",
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (require API access)",
]

[tool.uv.sources]
openhands-sdk = { git = "https://github.com/OpenHands/software-agent-sdk.git", subdirectory = "openhands-sdk" }
openhands-tools = { git = "https://github.com/OpenHands/software-agent-sdk.git", subdirectory = "openhands-tools" }

